% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjustAcuteAngles}
\alias{adjustAcuteAngles}
\title{Given a xy matrix of points, adjust the
points to avoid acute angles < 80 degrees}
\usage{
adjustAcuteAngles(xy, angle, minAngle = 80)
}
\arguments{
\item{xy}{xy dataframe}

\item{angle}{angle of the flight lines}

\item{minAngle}{the minimum angle to below which will be projected}
}
\description{
Given a xy matrix of points, adjust the
points to avoid acute angles < 80 degrees
}
