% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleBoundary}
\alias{exampleBoundary}
\title{Example boundary data}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 12 columns.}
\usage{
exampleBoundary
}
\description{
Example boundary data
}
\keyword{datasets}
