% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flight.summary}
\alias{flight.summary}
\title{Function to print flight plan summary}
\usage{
flight.summary(flight.params)
}
\arguments{
\item{flight.params}{Flight Parameters. parameters calculated from flight.parameters()}
}
\description{
This function will print messages with flight plan parameters,
like flight time, photo interval
}
\examples{

flight.params = flightplanning::flight.parameters(
  gsd = 4,
  side.overlap = 0.8,
  front.overlap = 0.8,
  flight.speed.kmh = 54
)
plansummary = flight.summary(flight.params)

}
