% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{flightParameters}
\alias{flightParameters}
\title{Function to generate Litchi csv flight plan}
\usage{
flightParameters(uav = "p3", GSD = 4, overlapWidth = 0.8,
  overlapHeight = 0.8, flightSpeedKmH = 30)
}
\arguments{
\item{uav}{either "p3" or "p4adv" for loading Phantom 3-4std or Phanton4-adv/pro
camera profiles}

\item{GSD}{target ground resolution in centimeters}

\item{overlapWidth}{desired width overlap between photos}

\item{overlapHeight}{desired height overlap between photos}

\item{flightSpeedKmH}{flight speed in km/h}
}
\description{
Function to generate Litchi csv flight plan
}
\examples{
params = flightParameters(
  GSD=4.325,
  flightSpeedKmH=30,
  overlapWidth = 0.8,
  overlapHeight = 0.8
 )

}
