% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateLitchiPlan}
\alias{generateLitchiPlan}
\title{Function to generate Litchi csv flight plan}
\usage{
generateLitchiPlan(ogrROI, outputPath, uav = "p3", GSD = 4,
  flightSpeedKmH = 30, overlapWidth = 0.8, overlapHeight = 0.8,
  gimbalPitchAngle = -90, flightLinesAngle = -1,
  maxWaypointsDistance = 2000)
}
\arguments{
\item{ogrROI}{range of interest loaded as a OGR layer, must be in
a metric units projection for working properly}

\item{outputPath}{output path for the csv file}

\item{uav}{either "p3" or "p4adv" for loading Phantom 3-4std or Phanton4-adv/pro
camera profiles, default "p3"}

\item{GSD}{target ground resolution in centimeters, default 4}

\item{flightSpeedKmH}{flight speed in km/h, default 30}

\item{overlapWidth}{desired width overlap between photos, default 0.8}

\item{overlapHeight}{desired height overlap between photos, defualt 0.8}

\item{gimbalPitchAngle}{gimbal angle for taking photos, default -90 (can be overriden at flight time)}

\item{flightLinesAngle}{angle for the flight lines, default -1 (auto set based on larger dimensions)}

\item{maxWaypointsDistance}{maximum distance between waypoints in meters,
default 2000 (it has been reported >2km will give problems)}
}
\description{
Function to generate Litchi csv flight plan
}
\examples{
data(exampleBoundary)
outPath = tempfile(fileext=".csv")
generateLitchiPlan(ogrROI = exampleBoundary,
                   outputPath = outPath,
                   uav = "p3",
                   GSD = 4.325,
                   flightSpeedKmH = 30,
                   overlapWidth = 0.8,
                   overlapHeight = 0.8,
                   gimbalPitchAngle = -90,
                   flightLinesAngle = -1,
                   maxWaypointsDistance = 2000)


}
