% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBBoxAngle}
\alias{getBBoxAngle}
\title{Provided an angle, calculate the corresponding minimum bounding box}
\usage{
getBBoxAngle(vertices, alpha)
}
\description{
Provided an angle, calculate the corresponding minimum bounding box
}
