% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litchi}
\alias{litchi}
\title{Litchi base csv data}
\format{An object of class \code{data.frame} with 1 rows and 45 columns.}
\usage{
litchi
}
\description{
Litchi base csv data
}
\keyword{datasets}
