% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjustAcuteAngles}
\alias{adjustAcuteAngles}
\title{Given a xy matrix of points, adjust the
points to avoid acute angles < 80 degrees}
\usage{
adjustAcuteAngles(xy, angle, minAngle)
}
\description{
Given a xy matrix of points, adjust the
points to avoid acute angles < 80 degrees
}
