% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{flight.parameters}
\alias{flight.parameters}
\title{Function to calculate flight parameters}
\usage{
flight.parameters(sensor.width = 6.17, focal.length35 = 20,
  aspect.ratio = "4:3", image.width.px = 4000, gsd = NA,
  side.overlap = 0.8, front.overlap = 0.8, flight.speed.kmh = NA,
  height = NA)
}
\arguments{
\item{sensor.width}{numeric. Camera sensor width in milimeters, default 6.17}

\item{focal.length35}{numeric. Camera focal length 35mm equivalent, default 20}

\item{aspect.ratio}{character. Aspect ratio of the picture, default "4:3"}

\item{image.width.px}{numeric. Width of the image in number of pixels, default 4000}

\item{gsd}{target ground resolution in centimeters, must provide either `gsd` or `height`}

\item{side.overlap}{desired width overlap between photos}

\item{front.overlap}{desired height overlap between photos}

\item{flight.speed.kmh}{flight speed in km/h}

\item{height}{target flight height, default NA}
}
\description{
This function will calculate the flight parameters by providing the camera settings
target flight height or gsd, front and side overlap.
}
\examples{
params = flight.parameters(
  gsd=4.325,
  flight.speed.kmh=30,
  side.overlap = 0.8,
  front.overlap = 0.8
 )

}
