% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{flightParameters}
\alias{flightParameters}
\title{Function to calculate flight parameters}
\usage{
flightParameters(sensorWidth = 6.17, focalLength35 = 20,
  aspectRatio = "4:3", imageWidthPx = 4000, GSD = 4,
  sideOverlap = 0.8, frontOverlap = 0.8, flightSpeedKmH = NA,
  flightHeight = NA)
}
\arguments{
\item{sensorWidth}{numeric. Camera sensor width in milimeters, default 6.17}

\item{focalLength35}{numeric. Camera focal length 35mm equivalent, default 20}

\item{aspectRatio}{character. Aspect ratio of the picture, default "4:3"}

\item{GSD}{target ground resolution in centimeters}

\item{sideOverlap}{desired width overlap between photos}

\item{frontOverlap}{desired height overlap between photos}

\item{flightSpeedKmH}{flight speed in km/h}

\item{imageWidth}{numeric. Width of the image in number of pixels, default 4000}
}
\description{
This function will calculate the flight parameters by providing the camera settings
  target flight height or GSD, front and side overlap.
}
\examples{
params = flightParameters(
  GSD=4.325,
  flightSpeedKmH=30,
  sideOverlap = 0.8,
  frontOverlap = 0.8
 )

}
