% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAngles}
\alias{getAngles}
\title{Get angles for each point considering
the two neighbors points}
\usage{
getAngles(waypoints)
}
\description{
Get angles for each point considering
the two neighbors points
}
