% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outerCurvePoints}
\alias{outerCurvePoints}
\title{Create outer curves for the flight lines}
\usage{
outerCurvePoints(waypoints, angle, flightLineDistance)
}
\description{
Create outer curves for the flight lines
}
